refactor(tui): extract DateBreakdownPanel, consolidate colors, fix date parsing
- Extract DateBreakdownPanel component to separate file (72 lines)
- Move SOURCE_COLORS to utils/colors.ts with SourceType typing
- Add getSourceColor() helper function
- Fix date parsing timezone drift with T00:00:00 suffix
- Add model sorting by total tokens in DateBreakdownPanel
- Remove debug code (debugInfo signal, debug UI, comments)
- Update component barrel exports

StatsView.tsx reduced from 316 to 225 lines.
