+++ packages/cli/src/tui/components/index.ts
@@ -11,3 +11,5 @@ export { LoadingSpinner } from "./LoadingSpinner.js";
 export { ModelRow } from "./ModelRow.js";
 export { TokenBreakdown } from "./TokenBreakdown.js";
 export type { TokenBreakdownData } from "./TokenBreakdown.js";
+export { DateBreakdownPanel } from "./DateBreakdownPanel.js";
+export type { DateBreakdownPanelProps } from "./DateBreakdownPanel.js";
